test: consolidate embedded agent and channel suites
