+++ extensions/telegram/src/bot.test.ts
@@ -100,7 +100,6 @@ const {
   replySpy,
   resolveExecApprovalSpy,
   sendMessageSpy,
-  setMyCommandsSpy,
   telegramBotDepsForTest,
   wasSentByBot,
 } = await import("./bot.create-telegram-bot.test-harness.js");
@@ -125,10 +124,42 @@ const PARTY_EMOJI = "\u{1F389}";
 const EYES_EMOJI = "\u{1F440}";
 const HEART_EMOJI = "\u{2764}\u{FE0F}";
 
+type TelegramChannelConfig = NonNullable<NonNullable<OpenClawConfig["channels"]>["telegram"]>;
+
+function mockTelegramConfig(
+  telegram: TelegramChannelConfig,
+  config: Omit<OpenClawConfig, "channels"> = {},
+) {
+  loadConfig.mockReturnValue({ ...config, channels: { telegram } } satisfies OpenClawConfig);
+}
+
+type TelegramCallbackQueryOverrides = {
+  id: string;
+  data: string;
+  from?: Record<string, unknown>;
+  message?: Record<string, unknown>;
+};
+
+type TelegramApprovalResolution = Awaited<ReturnType<typeof resolveExecApprovalSpy>>;
+
+function makeCallbackQuery(overrides: TelegramCallbackQueryOverrides) {
+  return {
+    id: overrides.id,
+    data: overrides.data,
+    from: overrides.from ?? { id: 9, first_name: "Ada", username: "ada_bot" },
+    message: {
+      chat: { id: 1234, type: "private" },
+      date: 1_736_380_800,
+      message_id: 10,
+      ...overrides.message,
+    },
+  };
+}
+
 type TelegramReactionPolicyCase = {
   name: string;
   updateId: number;
-  channelConfig: NonNullable<NonNullable<OpenClawConfig["channels"]>["telegram"]>;
+  channelConfig: TelegramChannelConfig;
   reaction?: Record<string, unknown>;
   sentByBot?: boolean;
   expectedEnqueueCalls: number;
@@ -632,19 +663,18 @@ describe("createTelegramBot", () => {
   });
 
   beforeEach(() => {
-    setMyCommandsSpy.mockClear();
-    questionGatewayHoisted.resolveQuestionOverGatewaySpy.mockClear();
+    vi.clearAllMocks();
     clearPluginInteractiveHandlers();
-    loadConfig.mockReturnValue({
-      agents: {
-        defaults: {
-          userTimezone: "UTC",
+    mockTelegramConfig(
+      { dmPolicy: "open", allowFrom: ["*"] },
+      {
+        agents: {
+          defaults: {
+            userTimezone: "UTC",
+          },
         },
       },
-      channels: {
-        telegram: { dmPolicy: "open", allowFrom: ["*"] },
-      },
-    });
+    );
     createTelegramBot = (opts) => {
       const telegramDeps = {
         ...telegramBotDepsForTest,
@@ -659,14 +689,9 @@ describe("createTelegramBot", () => {
   });
 
   it("starts with retired includeGroupHistoryContext still present in raw config", async () => {
-    loadConfig.mockReturnValue({
+    mockTelegramConfig({ includeGroupHistoryContext: "mention-only" } as never, {
       messages: { groupChat: { unmentionedInbound: "room_event" } },
-      channels: {
-        telegram: {
-          includeGroupHistoryContext: "mention-only",
-        },
-      },
-    } as never);
+    });
 
     createTelegramBot({ token: "tok" });
 
@@ -674,8 +699,6 @@ describe("createTelegramBot", () => {
   });
 
   it("dedupes outbound prompt-context sends with ambient group history", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
     const cfg = {
       messages: { groupChat: { unmentionedInbound: "room_event", mentionPatterns: [] } },
       channels: {
@@ -825,8 +848,6 @@ describe("createTelegramBot", () => {
   ])(
     "$caseName",
     async ({ replyFrom, expectedSender, omitMe, senderBusinessBot, chatId, replyMessageId }) => {
-      onSpy.mockClear();
-      replySpy.mockClear();
       const storePath = createTelegramTestStorePath(`self-projection-${chatId}`);
       const cfg = {
         channels: {
@@ -933,11 +954,6 @@ describe("createTelegramBot", () => {
   );
 
   it("uses the live allowlist when authorizing callbacks", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-    sendMessageSpy.mockClear();
-    loadConfig.mockClear();
-
     const startupConfig = {
       channels: {
         telegram: {
@@ -977,10 +993,6 @@ describe("createTelegramBot", () => {
   });
 
   it("blocks DM model-selection callbacks for unpaired users when inline buttons are DM-scoped", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-    editMessageTextSpy.mockClear();
-
     const storePath = createTelegramTestStorePath("callback-authz");
 
     const config = {
@@ -1029,10 +1041,6 @@ describe("createTelegramBot", () => {
   });
 
   it("blocks group model-selection callbacks for senders who are not authorized for /models", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-    editMessageTextSpy.mockClear();
-
     const storePath = createTelegramTestStorePath("group-model-authz");
 
     const config = {
@@ -1089,10 +1097,6 @@ describe("createTelegramBot", () => {
   });
 
   it("recomputes group model-selection callback auth from runtime command config", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-    editMessageTextSpy.mockClear();
-
     const storePath = createTelegramTestStorePath("group-model-authz-runtime");
 
     try {
@@ -1158,24 +1162,14 @@ describe("createTelegramBot", () => {
       expect(answerCallbackQuerySpy).toHaveBeenCalledWith("cbq-group-model-authz-runtime-1");
     } finally {
       loadConfig.mockReset();
-      loadConfig.mockReturnValue({
-        agents: {
-          defaults: {
-            envelopeTimezone: "utc",
-          },
-        },
-        channels: {
-          telegram: { dmPolicy: "open", allowFrom: ["*"] },
-        },
-      });
+      mockTelegramConfig(
+        { dmPolicy: "open", allowFrom: ["*"] },
+        { agents: { defaults: { envelopeTimezone: "utc" } } },
+      );
     }
   });
 
   it("allows callback_query in groups when group policy authorizes the sender", async () => {
-    onSpy.mockClear();
-    editMessageTextSpy.mockClear();
-    listSkillCommandsForAgents.mockClear();
-
     createTelegramBot({
       token: "tok",
       config: {
@@ -1210,9 +1204,6 @@ describe("createTelegramBot", () => {
   });
 
   it("keeps group question callbacks on the configured callback allowlist", async () => {
-    onSpy.mockClear();
-    answerCallbackQuerySpy.mockClear();
-
     const config = {
       channels: {
         telegram: {
@@ -1245,23 +1236,10 @@ describe("createTelegramBot", () => {
   });
 
   it("replaces legacy approval controls with a visible terminal receipt", async () => {
-    onSpy.mockClear();
-    editMessageReplyMarkupSpy.mockClear();
-    editMessageTextSpy.mockClear();
-    resolveExecApprovalSpy.mockClear();
-
-    loadConfig.mockReturnValue({
-      channels: {
-        telegram: {
-          dmPolicy: "open",
-          allowFrom: ["*"],
-          execApprovals: {
-            enabled: true,
-            approvers: ["9"],
-            target: "dm",
-          },
-        },
-      },
+    mockTelegramConfig({
+      dmPolicy: "open",
+      allowFrom: ["*"],
+      execApprovals: { enabled: true, approvers: ["9"], target: "dm" },
     });
     createTelegramBot({ token: "tok" });
     const callbackHandler = getTelegramCallbackHandlerForTests();
@@ -1318,24 +1296,12 @@ describe("createTelegramBot", () => {
   });
 
   it("allows approval callbacks when exec approvals are enabled even without generic inlineButtons capability", async () => {
-    onSpy.mockClear();
-    editMessageReplyMarkupSpy.mockClear();
-    editMessageTextSpy.mockClear();
-
-    loadConfig.mockReturnValue({
-      channels: {
-        telegram: {
-          botToken: "tok",
-          dmPolicy: "open",
-          allowFrom: ["*"],
-          capabilities: ["vision"],
-          execApprovals: {
-            enabled: true,
-            approvers: ["9"],
-            target: "dm",
-          },
-        },
-      },
+    mockTelegramConfig({
+      botToken: "tok",
+      dmPolicy: "open",
+      allowFrom: ["*"],
+      capabilities: ["vision"],
+      execApprovals: { enabled: true, approvers: ["9"], target: "dm" },
     });
     createTelegramBot({ token: "tok" });
     const callbackHandler = getTelegramCallbackHandlerForTests();
@@ -1354,10 +1320,6 @@ describe("createTelegramBot", () => {
   });
 
   it("uses explicit ownership and renders canonical truth on a losing typed surface", async () => {
-    onSpy.mockClear();
-    editMessageReplyMarkupSpy.mockClear();
-    editMessageTextSpy.mockClear();
-    resolveExecApprovalSpy.mockClear();
     resolveExecApprovalSpy.mockResolvedValueOnce({
       applied: false,
       approval: {
@@ -1378,18 +1340,10 @@ describe("createTelegramBot", () => {
       },
     });
 
-    loadConfig.mockReturnValue({
-      channels: {
-        telegram: {
-          dmPolicy: "open",
-          allowFrom: ["*"],
-          execApprovals: {
-            enabled: true,
-            approvers: ["9"],
-            target: "dm",
-          },
-        },
-      },
+    mockTelegramConfig({
+      dmPolicy: "open",
+      allowFrom: ["*"],
+      execApprovals: { enabled: true, approvers: ["9"], target: "dm" },
     });
     createTelegramBot({ token: "tok" });
     const callbackHandler = getTelegramCallbackHandlerForTests();
@@ -1435,11 +1389,6 @@ describe("createTelegramBot", () => {
   });
 
   it("sends a canonical terminal receipt when the clicked approval message cannot be edited", async () => {
-    onSpy.mockClear();
-    editMessageReplyMarkupSpy.mockClear();
-    editMessageTextSpy.mockClear();
-    sendMessageSpy.mockClear();
-    resolveExecApprovalSpy.mockClear();
     editMessageTextSpy.mockRejectedValueOnce(new Error("Bad Request: message can't be edited"));
     resolveExecApprovalSpy.mockResolvedValueOnce({
       applied: true,
@@ -1461,18 +1410,10 @@ describe("createTelegramBot", () => {
       },
     });
 
-    loadConfig.mockReturnValue({
-      channels: {
-        telegram: {
-          dmPolicy: "open",
-          allowFrom: ["*"],
-          execApprovals: {
-            enabled: true,
-            approvers: ["9"],
-            target: "dm",
-          },
-        },
-      },
+    mockTelegramConfig({
+      dmPolicy: "open",
+      allowFrom: ["*"],
+      execApprovals: { enabled: true, approvers: ["9"], target: "dm" },
     });
     createTelegramBot({ token: "tok" });
     const callbackData = buildTelegramApprovalCallbackData({
@@ -1486,17 +1427,14 @@ describe("createTelegramBot", () => {
     }
 
     await getTelegramCallbackHandlerForTests()({
-      callbackQuery: {
+      callbackQuery: makeCallbackQuery({
         id: "cbq-terminal-edit-fallback",
         data: callbackData,
-        from: { id: 9, first_name: "Ada", username: "ada_bot" },
         message: {
-          chat: { id: 1234, type: "private" },
-          date: 1736380800,
           message_id: 25,
           text: "Approval required.",
         },
-      },
+      }),
       me: { username: "openclaw_bot" },
       getFile: getEmptyTelegramFile,
     });
@@ -1520,12 +1458,6 @@ describe("createTelegramBot", () => {
   });
 
   it("consumes malformed callbacks in the reserved approval namespace", async () => {
-    onSpy.mockClear();
-    editMessageReplyMarkupSpy.mockClear();
-    editMessageTextSpy.mockClear();
-    enqueueSystemEventSpy.mockClear();
-    replySpy.mockClear();
-    resolveExecApprovalSpy.mockClear();
     const pluginHandler = vi.fn(async () => ({ handled: true }));
     registerPluginInteractiveHandler("reserved-approval-test", {
       channel: "telegram",
@@ -1533,33 +1465,22 @@ describe("createTelegramBot", () => {
       handler: pluginHandler as never,
     });
 
-    loadConfig.mockReturnValue({
-      channels: {
-        telegram: {
-          dmPolicy: "open",
-          allowFrom: ["*"],
-          execApprovals: {
-            enabled: true,
-            approvers: ["9"],
-            target: "dm",
-          },
-        },
-      },
+    mockTelegramConfig({
+      dmPolicy: "open",
+      allowFrom: ["*"],
+      execApprovals: { enabled: true, approvers: ["9"], target: "dm" },
     });
     createTelegramBot({ token: "tok" });
 
     await getTelegramCallbackHandlerForTests()({
-      callbackQuery: {
+      callbackQuery: makeCallbackQuery({
         id: "cbq-malformed-reserved-approval",
         data: "tga1:e:x:req-1",
-        from: { id: 9, first_name: "Ada", username: "ada_bot" },
         message: {
-          chat: { id: 1234, type: "private" },
-          date: 1736380800,
           message_id: 26,
           text: "Approval required.",
         },
-      },
+      }),
       me: { username: "openclaw_bot" },
       getFile: getEmptyTelegramFile,
     });
@@ -1578,145 +1499,113 @@ describe("createTelegramBot", () => {
     expect(answerCallbackQuerySpy).toHaveBeenCalledWith("cbq-malformed-reserved-approval");
   });
 
-  it("terminalizes a stale legacy click from the canonical record without retrying owners", async () => {
-    onSpy.mockClear();
-    editMessageReplyMarkupSpy.mockClear();
-    editMessageTextSpy.mockClear();
-    resolveExecApprovalSpy.mockClear();
-    const alreadyResolved = Object.assign(new Error("approval already resolved"), {
-      gatewayCode: "INVALID_REQUEST",
-      details: { reason: "APPROVAL_ALREADY_RESOLVED" },
-    });
-    resolveExecApprovalSpy.mockRejectedValueOnce(alreadyResolved).mockResolvedValueOnce({
-      applied: false,
-      approval: {
-        id: "stale-legacy-id",
-        urlPath: "/approve/stale-legacy-id",
-        createdAtMs: 1,
-        expiresAtMs: 60_000,
-        resolvedAtMs: 2,
-        reason: "user",
-        status: "denied",
-        decision: "deny",
-        presentation: {
-          kind: "exec",
-          commandText: "echo denied",
-          allowedDecisions: ["allow-once", "deny"],
-        },
-      },
-    });
-
-    loadConfig.mockReturnValue({
-      channels: {
-        telegram: {
-          dmPolicy: "open",
-          allowFrom: ["*"],
-          execApprovals: {
-            enabled: true,
-            approvers: ["9"],
-            target: "dm",
+  it.each([
+    {
+      name: "terminalizes a stale legacy click from the canonical record without retrying owners",
+      callbackId: "cbq-stale-legacy",
+      callbackData: "/approve stale-legacy-id allow-once",
+      messageId: 25,
+      secondResolution: {
+        kind: "resolved" as const,
+        value: {
+          applied: false,
+          approval: {
+            id: "stale-legacy-id",
+            urlPath: "/approve/stale-legacy-id",
+            createdAtMs: 1,
+            expiresAtMs: 60_000,
+            resolvedAtMs: 2,
+            reason: "user",
+            status: "denied",
+            decision: "deny",
+            presentation: {
+              kind: "exec",
+              commandText: "echo denied",
+              allowedDecisions: ["allow-once", "deny"],
+            },
           },
-        },
+        } satisfies TelegramApprovalResolution,
       },
-    });
-    createTelegramBot({ token: "tok" });
-
-    await getTelegramCallbackHandlerForTests()({
-      callbackQuery: {
-        id: "cbq-stale-legacy",
-        data: "/approve stale-legacy-id allow-once",
-        from: { id: 9, first_name: "Ada", username: "ada_bot" },
-        message: {
-          chat: { id: 1234, type: "private" },
-          date: 1736380800,
-          message_id: 25,
-          text: "Approval required.",
-        },
+      expectedTerminalText: "Canonical result: Denied",
+      assertDistinctResult: () => {
+        expect(execApprovalCall(0)).toMatchObject({
+          approvalId: "stale-legacy-id",
+          approvalKind: "exec",
+        });
+        expect(execApprovalCall(1)).toMatchObject({
+          approvalId: "stale-legacy-id",
+          approvalKind: "exec",
+        });
+        expect(answerCallbackQuerySpy).toHaveBeenCalledWith("cbq-stale-legacy");
       },
-      me: { username: "openclaw_bot" },
-      getFile: getEmptyTelegramFile,
-    });
-
-    expect(resolveExecApprovalSpy).toHaveBeenCalledTimes(2);
-    expect(execApprovalCall(0)).toMatchObject({
-      approvalId: "stale-legacy-id",
-      approvalKind: "exec",
-    });
-    expect(execApprovalCall(1)).toMatchObject({
-      approvalId: "stale-legacy-id",
-      approvalKind: "exec",
-    });
-    expect(editMessageTextSpy).toHaveBeenCalledWith(
-      1234,
-      25,
-      expect.stringContaining("Canonical result: Denied"),
-      { reply_markup: { inline_keyboard: [] } },
-    );
-    expect(editMessageReplyMarkupSpy).not.toHaveBeenCalled();
-    expect(answerCallbackQuerySpy).toHaveBeenCalledWith("cbq-stale-legacy");
-  });
-
-  it("renders neutral terminal copy when a stale legacy record cannot be fetched", async () => {
-    onSpy.mockClear();
-    editMessageReplyMarkupSpy.mockClear();
-    editMessageTextSpy.mockClear();
-    resolveExecApprovalSpy.mockClear();
-    const alreadyResolved = Object.assign(new Error("approval already resolved"), {
-      gatewayCode: "INVALID_REQUEST",
-      details: { reason: "APPROVAL_ALREADY_RESOLVED" },
-    });
-    resolveExecApprovalSpy
-      .mockRejectedValueOnce(alreadyResolved)
-      .mockRejectedValueOnce(new Error("unknown or expired approval id"));
-
-    loadConfig.mockReturnValue({
-      channels: {
-        telegram: {
-          dmPolicy: "open",
-          allowFrom: ["*"],
-          execApprovals: {
-            enabled: true,
-            approvers: ["9"],
-            target: "dm",
-          },
-        },
+    },
+    {
+      name: "renders neutral terminal copy when a stale legacy record cannot be fetched",
+      callbackId: "cbq-stale-legacy-neutral",
+      callbackData: "/approve stale-neutral-id deny",
+      messageId: 26,
+      secondResolution: {
+        kind: "rejected" as const,
+        value: new Error("unknown or expired approval id"),
       },
-    });
-    createTelegramBot({ token: "tok" });
-
-    await getTelegramCallbackHandlerForTests()({
-      callbackQuery: {
-        id: "cbq-stale-legacy-neutral",
-        data: "/approve stale-neutral-id deny",
-        from: { id: 9, first_name: "Ada", username: "ada_bot" },
-        message: {
-          chat: { id: 1234, type: "private" },
-          date: 1736380800,
-          message_id: 26,
-          text: "Approval required.",
-        },
-      },
-      me: { username: "openclaw_bot" },
-      getFile: getEmptyTelegramFile,
-    });
-
-    expect(resolveExecApprovalSpy).toHaveBeenCalledTimes(2);
-    expect(editMessageTextSpy).toHaveBeenCalledWith(
-      1234,
-      26,
-      expect.stringContaining(
+      expectedTerminalText:
         "It was already resolved or expired; the canonical decision is unavailable here.",
-      ),
-      { reply_markup: { inline_keyboard: [] } },
-    );
-    expect(editMessageReplyMarkupSpy).not.toHaveBeenCalled();
-  });
+      assertDistinctResult: undefined,
+    },
+  ])(
+    "$name",
+    async ({
+      callbackId,
+      callbackData,
+      messageId,
+      secondResolution,
+      expectedTerminalText,
+      assertDistinctResult,
+    }) => {
+      const alreadyResolved = Object.assign(new Error("approval already resolved"), {
+        gatewayCode: "INVALID_REQUEST",
+        details: { reason: "APPROVAL_ALREADY_RESOLVED" },
+      });
+      resolveExecApprovalSpy.mockRejectedValueOnce(alreadyResolved);
+      if (secondResolution.kind === "resolved") {
+        resolveExecApprovalSpy.mockResolvedValueOnce(secondResolution.value);
+      } else {
+        resolveExecApprovalSpy.mockRejectedValueOnce(secondResolution.value);
+      }
+
+      mockTelegramConfig({
+        dmPolicy: "open",
+        allowFrom: ["*"],
+        execApprovals: { enabled: true, approvers: ["9"], target: "dm" },
+      });
+      createTelegramBot({ token: "tok" });
+
+      await getTelegramCallbackHandlerForTests()({
+        callbackQuery: makeCallbackQuery({
+          id: callbackId,
+          data: callbackData,
+          message: {
+            message_id: messageId,
+            text: "Approval required.",
+          },
+        }),
+        me: { username: "openclaw_bot" },
+        getFile: getEmptyTelegramFile,
+      });
+
+      expect(resolveExecApprovalSpy).toHaveBeenCalledTimes(2);
+      expect(editMessageTextSpy).toHaveBeenCalledWith(
+        1234,
+        messageId,
+        expect.stringContaining(expectedTerminalText),
+        { reply_markup: { inline_keyboard: [] } },
+      );
+      expect(editMessageReplyMarkupSpy).not.toHaveBeenCalled();
+      assertDistinctResult?.();
+    },
+  );
 
   it("retries a stale legacy click when canonical convergence fails transiently", async () => {
-    onSpy.mockClear();
-    editMessageReplyMarkupSpy.mockClear();
-    editMessageTextSpy.mockClear();
-    resolveExecApprovalSpy.mockClear();
     const alreadyResolved = Object.assign(new Error("approval already resolved"), {
       gatewayCode: "INVALID_REQUEST",
       details: { reason: "APPROVAL_ALREADY_RESOLVED" },
@@ -1725,34 +1614,23 @@ describe("createTelegramBot", () => {
       .mockRejectedValueOnce(alreadyResolved)
       .mockRejectedValueOnce(new Error("gateway unavailable"));
 
-    loadConfig.mockReturnValue({
-      channels: {
-        telegram: {
-          dmPolicy: "open",
-          allowFrom: ["*"],
-          execApprovals: {
-            enabled: true,
-            approvers: ["9"],
-            target: "dm",
-          },
-        },
-      },
+    mockTelegramConfig({
+      dmPolicy: "open",
+      allowFrom: ["*"],
+      execApprovals: { enabled: true, approvers: ["9"], target: "dm" },
     });
     createTelegramBot({ token: "tok" });
 
     await expect(
       getTelegramCallbackHandlerForTests()({
-        callbackQuery: {
+        callbackQuery: makeCallbackQuery({
           id: "cbq-stale-legacy-retry",
           data: "/approve stale-retry-id deny",
-          from: { id: 9, first_name: "Ada", username: "ada_bot" },
           message: {
-            chat: { id: 1234, type: "private" },
-            date: 1736380800,
             message_id: 27,
             text: "Approval required.",
           },
-        },
+        }),
         me: { username: "openclaw_bot" },
         getFile: getEmptyTelegramFile,
       }),
@@ -1765,24 +1643,12 @@ describe("createTelegramBot", () => {
   });
 
   it("resolves legacy opaque plugin ids without inferring kind from id spelling", async () => {
-    onSpy.mockClear();
-    editMessageReplyMarkupSpy.mockClear();
-    editMessageTextSpy.mockClear();
-    resolveExecApprovalSpy.mockClear();
     resolveExecApprovalSpy.mockRejectedValueOnce(new Error("unknown or expired approval id"));
 
-    loadConfig.mockReturnValue({
-      channels: {
-        telegram: {
-          dmPolicy: "open",
-          allowFrom: ["*"],
-          execApprovals: {
-            enabled: true,
-            approvers: ["9"],
-            target: "dm",
-          },
-        },
-      },
+    mockTelegramConfig({
+      dmPolicy: "open",
+      allowFrom: ["*"],
+      execApprovals: { enabled: true, approvers: ["9"], target: "dm" },
     });
     createTelegramBot({ token: "tok" });
     const callbackHandler = getTelegramCallbackHandlerForTests();
@@ -1825,22 +1691,10 @@ describe("createTelegramBot", () => {
   });
 
   it("does not resolve opaque approval-shaped plugin callbacks", async () => {
-    onSpy.mockClear();
-    editMessageReplyMarkupSpy.mockClear();
-    resolveExecApprovalSpy.mockClear();
-
-    loadConfig.mockReturnValue({
-      channels: {
-        telegram: {
-          dmPolicy: "open",
-          allowFrom: ["*"],
-          execApprovals: {
-            enabled: true,
-            approvers: ["9"],
-            target: "dm",
-          },
-        },
-      },
+    mockTelegramConfig({
+      dmPolicy: "open",
+      allowFrom: ["*"],
+      execApprovals: { enabled: true, approvers: ["9"], target: "dm" },
     });
     createTelegramBot({ token: "tok" });
     const callbackHandler = getTelegramCallbackHandlerForTests();
@@ -1859,23 +1713,10 @@ describe("createTelegramBot", () => {
   });
 
   it("blocks approval callbacks from telegram users who are not exec approvers", async () => {
-    onSpy.mockClear();
-    editMessageReplyMarkupSpy.mockClear();
-    editMessageTextSpy.mockClear();
-    resolveExecApprovalSpy.mockClear();
-
-    loadConfig.mockReturnValue({
-      channels: {
-        telegram: {
-          dmPolicy: "open",
-          allowFrom: ["*"],
-          execApprovals: {
-            enabled: true,
-            approvers: ["999"],
-            target: "dm",
-          },
-        },
-      },
+    mockTelegramConfig({
+      dmPolicy: "open",
+      allowFrom: ["*"],
+      execApprovals: { enabled: true, approvers: ["999"], target: "dm" },
     });
     createTelegramBot({ token: "tok" });
     const callbackHandler = getTelegramCallbackHandlerForTests();
@@ -1895,24 +1736,12 @@ describe("createTelegramBot", () => {
   });
 
   it("keeps approval callback resolution failures out of Telegram chat before retry", async () => {
-    onSpy.mockClear();
-    sendMessageSpy.mockClear();
-    editMessageReplyMarkupSpy.mockClear();
-    resolveExecApprovalSpy.mockClear();
     resolveExecApprovalSpy.mockRejectedValueOnce(new Error("gateway secret detail"));
 
-    loadConfig.mockReturnValue({
-      channels: {
-        telegram: {
-          dmPolicy: "open",
-          allowFrom: ["*"],
-          execApprovals: {
-            enabled: true,
-            approvers: ["9"],
-            target: "dm",
-          },
-        },
-      },
+    mockTelegramConfig({
+      dmPolicy: "open",
+      allowFrom: ["*"],
+      execApprovals: { enabled: true, approvers: ["9"], target: "dm" },
     });
     createTelegramBot({ token: "tok" });
     const callbackHandler = getTelegramCallbackHandlerForTests();
@@ -1933,26 +1762,18 @@ describe("createTelegramBot", () => {
   });
 
   it("allows target-only exec resolution despite a misleading plugin id prefix", async () => {
-    onSpy.mockClear();
-    editMessageReplyMarkupSpy.mockClear();
-    editMessageTextSpy.mockClear();
-    resolveExecApprovalSpy.mockClear();
-
-    loadConfig.mockReturnValue({
-      approvals: {
-        exec: {
-          enabled: true,
-          mode: "targets",
-          targets: [{ channel: "telegram", to: "9" }],
+    mockTelegramConfig(
+      { dmPolicy: "open", allowFrom: ["*"] },
+      {
+        approvals: {
+          exec: {
+            enabled: true,
+            mode: "targets",
+            targets: [{ channel: "telegram", to: "9" }],
+          },
         },
       },
-      channels: {
-        telegram: {
-          dmPolicy: "open",
-          allowFrom: ["*"],
-        },
-      },
-    });
+    );
     createTelegramBot({ token: "tok" });
     const callbackHandler = getTelegramCallbackHandlerForTests();
 
@@ -1984,29 +1805,20 @@ describe("createTelegramBot", () => {
   });
 
   it("preserves ambiguous target-only stale callbacks for another approver", async () => {
-    onSpy.mockClear();
-    editMessageReplyMarkupSpy.mockClear();
-    editMessageTextSpy.mockClear();
-    resolveExecApprovalSpy.mockClear();
-    replySpy.mockClear();
-    sendMessageSpy.mockClear();
     resolveExecApprovalSpy.mockRejectedValueOnce(new Error("unknown or expired approval id"));
 
-    loadConfig.mockReturnValue({
-      approvals: {
-        exec: {
-          enabled: true,
-          mode: "targets",
-          targets: [{ channel: "telegram", to: "9" }],
+    mockTelegramConfig(
+      { dmPolicy: "open", allowFrom: ["*"] },
+      {
+        approvals: {
+          exec: {
+            enabled: true,
+            mode: "targets",
+            targets: [{ channel: "telegram", to: "9" }],
+          },
         },
       },
-      channels: {
-        telegram: {
-          dmPolicy: "open",
-          allowFrom: ["*"],
-        },
-      },
-    });
+    );
     createTelegramBot({ token: "tok" });
     const callbackHandler = getTelegramCallbackHandlerForTests();
 
@@ -2035,28 +1847,14 @@ describe("createTelegramBot", () => {
   });
 
   it("renders a terminal no-longer-pending receipt for expired legacy callbacks", async () => {
-    onSpy.mockClear();
-    editMessageReplyMarkupSpy.mockClear();
-    editMessageTextSpy.mockClear();
-    resolveExecApprovalSpy.mockClear();
-    replySpy.mockClear();
-    sendMessageSpy.mockClear();
     resolveExecApprovalSpy
       .mockRejectedValueOnce(new Error("unknown or expired approval id"))
       .mockRejectedValueOnce(new Error("unknown or expired approval id"));
 
-    loadConfig.mockReturnValue({
-      channels: {
-        telegram: {
-          dmPolicy: "open",
-          allowFrom: ["*"],
-          execApprovals: {
-            enabled: true,
-            approvers: ["9"],
-            target: "dm",
-          },
-        },
-      },
+    mockTelegramConfig({
+      dmPolicy: "open",
+      allowFrom: ["*"],
+      execApprovals: { enabled: true, approvers: ["9"], target: "dm" },
     });
     createTelegramBot({ token: "tok" });
     const callbackHandler = getTelegramCallbackHandlerForTests();
@@ -2089,43 +1887,32 @@ describe("createTelegramBot", () => {
   });
 
   it("does not call canonical resolution with a guessed kind after a legacy miss", async () => {
-    onSpy.mockClear();
-    editMessageReplyMarkupSpy.mockClear();
-    editMessageTextSpy.mockClear();
-    resolveExecApprovalSpy.mockClear();
     resolveExecApprovalSpy.mockRejectedValueOnce(new Error("unknown or expired approval id"));
 
-    loadConfig.mockReturnValue({
-      approvals: {
-        exec: {
-          enabled: true,
-          mode: "targets",
-          targets: [{ channel: "telegram", to: "9" }],
+    mockTelegramConfig(
+      { dmPolicy: "open", allowFrom: ["*"], capabilities: ["vision"] },
+      {
+        approvals: {
+          exec: {
+            enabled: true,
+            mode: "targets",
+            targets: [{ channel: "telegram", to: "9" }],
+          },
         },
       },
-      channels: {
-        telegram: {
-          dmPolicy: "open",
-          allowFrom: ["*"],
-          capabilities: ["vision"],
-        },
-      },
-    });
+    );
     createTelegramBot({ token: "tok" });
     const callbackHandler = getTelegramCallbackHandlerForTests();
 
     await callbackHandler({
-      callbackQuery: {
+      callbackQuery: makeCallbackQuery({
         id: "cbq-plugin-approve-blocked",
         data: "/approve plugin:138e9b8c allow-once",
-        from: { id: 9, first_name: "Ada", username: "ada_bot" },
         message: {
-          chat: { id: 1234, type: "private" },
-          date: 1736380800,
           message_id: 24,
           text: "Plugin approval required.",
         },
-      },
+      }),
       me: { username: "openclaw_bot" },
       getFile: getEmptyTelegramFile,
     });
@@ -2156,8 +1943,6 @@ describe("createTelegramBot", () => {
       messageId: 14,
     },
   ])("$name", async ({ callbackId, callbackData, messageId }) => {
-    onSpy.mockClear();
-    listSkillCommandsForAgents.mockClear();
     (
       listSkillCommandsForAgents as unknown as {
         mockImplementationOnce(implementation: TelegramBotDeps["listSkillCommandsForAgents"]): void;
@@ -2168,7 +1953,6 @@ describe("createTelegramBot", () => {
       }
       return [];
     });
-    editMessageTextSpy.mockClear();
 
     createTelegramBot({ token: "tok" });
     const callbackHandler = getTelegramCallbackHandlerForTests();
@@ -2204,10 +1988,6 @@ describe("createTelegramBot", () => {
   });
 
   it("ignores unsafe command pagination pages", async () => {
-    onSpy.mockClear();
-    listSkillCommandsForAgents.mockClear();
-    editMessageTextSpy.mockClear();
-
     createTelegramBot({ token: "tok" });
     const callbackHandler = getTelegramCallbackHandlerForTests();
 
@@ -2224,9 +2004,6 @@ describe("createTelegramBot", () => {
   });
 
   it("blocks pagination callbacks when allowlist rejects sender", async () => {
-    onSpy.mockClear();
-    editMessageTextSpy.mockClear();
-
     const config = {
       channels: {
         telegram: {
@@ -2278,10 +2055,6 @@ describe("createTelegramBot", () => {
   ])(
     "$name",
     async ({ callbackId, callbackData, configuredModels, defaultModel, messageId, storeLabel }) => {
-      onSpy.mockClear();
-      replySpy.mockClear();
-      editMessageTextSpy.mockClear();
-
       const storePath = createTelegramTestStorePath(storeLabel);
       const config: OpenClawConfig = {
         agents: {
@@ -2329,10 +2102,6 @@ describe("createTelegramBot", () => {
   );
 
   it("renders model callback lists with configured display names", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-    editMessageTextSpy.mockClear();
-
     const storePath = createTelegramTestStorePath("model-display-names");
     const buildModelsProviderDataMock =
       telegramBotDepsForTest.buildModelsProviderData as unknown as ReturnType<typeof vi.fn>;
@@ -2398,10 +2167,6 @@ describe("createTelegramBot", () => {
   });
 
   it("formats non-default model selection confirmations with Telegram HTML parse mode", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-    editMessageTextSpy.mockClear();
-
     const storePath = createTelegramTestStorePath("model-html");
 
     const config = {
@@ -2465,9 +2230,6 @@ describe("createTelegramBot", () => {
     // store path and default-model resolution.  If the config was reloaded
     // (e.g. default model changed) the override could be written to the wrong
     // store or incorrectly cleared because `isDefaultSelection` was wrong.
-    onSpy.mockClear();
-    replySpy.mockClear();
-    editMessageTextSpy.mockClear();
 
     const storePath = createTelegramTestStorePath("model-fresh-cfg");
     const debounceMs = 4321;
@@ -2612,10 +2374,6 @@ describe("createTelegramBot", () => {
   });
 
   it("rejects ambiguous compact model callbacks and returns provider list", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-    editMessageTextSpy.mockClear();
-
     createTelegramBot({
       token: "tok",
       config: {
@@ -2653,22 +2411,10 @@ describe("createTelegramBot", () => {
   });
 
   it("includes sender identity in group envelope headers", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-
-    loadConfig.mockReturnValue({
-      agents: {
-        defaults: {
-          userTimezone: "UTC",
-        },
-      },
-      channels: {
-        telegram: {
-          groupPolicy: "open",
-          groups: { "*": { requireMention: false } },
-        },
-      },
-    });
+    mockTelegramConfig(
+      { groupPolicy: "open", groups: { "*": { requireMention: false } } },
+      { agents: { defaults: { userTimezone: "UTC" } } },
+    );
 
     createTelegramBot({ token: "tok" });
     const handler = getOnHandler("message") as (ctx: Record<string, unknown>) => Promise<void>;
@@ -2707,22 +2453,10 @@ describe("createTelegramBot", () => {
   });
 
   it("adds live chat and reply-target windows for stale group replies", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-
-    loadConfig.mockReturnValue({
-      agents: {
-        defaults: {
-          envelopeTimezone: "utc",
-        },
-      },
-      channels: {
-        telegram: {
-          groupPolicy: "open",
-          groups: { "*": { requireMention: false } },
-        },
-      },
-    });
+    mockTelegramConfig(
+      { groupPolicy: "open", groups: { "*": { requireMention: false } } },
+      { agents: { defaults: { envelopeTimezone: "utc" } } },
+    );
 
     createTelegramBot({ token: "tok" });
     const handler = getOnHandler("message") as (ctx: Record<string, unknown>) => Promise<void>;
@@ -2805,23 +2539,14 @@ describe("createTelegramBot", () => {
   });
 
   it("keeps skipped group messages in default recent group history context", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-
-    loadConfig.mockReturnValue({
-      agents: {
-        defaults: {
-          envelopeTimezone: "utc",
-        },
+    mockTelegramConfig(
+      {
+        groupPolicy: "allowlist",
+        groupAllowFrom: ["111", "222"],
+        groups: { "*": { requireMention: true } },
       },
-      channels: {
-        telegram: {
-          groupPolicy: "allowlist",
-          groupAllowFrom: ["111", "222"],
-          groups: { "*": { requireMention: true } },
-        },
-      },
-    });
+      { agents: { defaults: { envelopeTimezone: "utc" } } },
+    );
 
     createTelegramBot({ token: "tok" });
     const handler = getOnHandler("message") as (ctx: Record<string, unknown>) => Promise<void>;
@@ -2886,23 +2611,14 @@ describe("createTelegramBot", () => {
   });
 
   it("excludes ambient transcript rows from live group conversation context", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-
-    loadConfig.mockReturnValue({
-      agents: {
-        defaults: {
-          envelopeTimezone: "utc",
-        },
+    mockTelegramConfig(
+      {
+        groupPolicy: "allowlist",
+        groupAllowFrom: ["111", "222", "333", "444"],
+        groups: { "*": { requireMention: true } },
       },
-      channels: {
-        telegram: {
-          groupPolicy: "allowlist",
-          groupAllowFrom: ["111", "222", "333", "444"],
-          groups: { "*": { requireMention: true } },
-        },
-      },
-    });
+      { agents: { defaults: { envelopeTimezone: "utc" } } },
+    );
 
     const previousReadAmbient = telegramBotDepsForTest.readAmbientTranscriptWatermark;
     const previousResolveAmbientKey = telegramBotDepsForTest.resolveAmbientTranscriptWatermarkKey;
@@ -2990,24 +2706,15 @@ describe("createTelegramBot", () => {
   });
 
   it("honors historyLimit zero for group chat-window context", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-
-    loadConfig.mockReturnValue({
-      agents: {
-        defaults: {
-          envelopeTimezone: "utc",
-        },
+    mockTelegramConfig(
+      {
+        groupPolicy: "allowlist",
+        groupAllowFrom: ["111", "222"],
+        historyLimit: 0,
+        groups: { "*": { requireMention: true } },
       },
-      channels: {
-        telegram: {
-          groupPolicy: "allowlist",
-          groupAllowFrom: ["111", "222"],
-          historyLimit: 0,
-          groups: { "*": { requireMention: true } },
-        },
-      },
-    });
+      { agents: { defaults: { envelopeTimezone: "utc" } } },
+    );
 
     createTelegramBot({ token: "tok" });
     const handler = getOnHandler("message") as (ctx: Record<string, unknown>) => Promise<void>;
@@ -3047,22 +2754,10 @@ describe("createTelegramBot", () => {
   });
 
   it("updates cached bot messages from Telegram edit updates", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-
-    loadConfig.mockReturnValue({
-      agents: {
-        defaults: {
-          envelopeTimezone: "utc",
-        },
-      },
-      channels: {
-        telegram: {
-          groupPolicy: "open",
-          groups: { "*": { requireMention: false } },
-        },
-      },
-    });
+    mockTelegramConfig(
+      { groupPolicy: "open", groups: { "*": { requireMention: false } } },
+      { agents: { defaults: { envelopeTimezone: "utc" } } },
+    );
 
     createTelegramBot({ token: "tok" });
     const handler = getOnHandler("message") as (ctx: Record<string, unknown>) => Promise<void>;
@@ -3144,9 +2839,6 @@ describe("createTelegramBot", () => {
   });
 
   it("keeps direct Telegram media context when transcript context exists", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-
     const storePath = createTelegramTestStorePath("dm-media-context");
     const config = {
       channels: {
@@ -3220,9 +2912,6 @@ describe("createTelegramBot", () => {
   });
 
   it("dedupes an unversioned outbound row by its explicit legacy prompt timestamp", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-
     const storePath = createTelegramTestStorePath("dm-legacy-dedupe");
     const config = {
       channels: { telegram: { dmPolicy: "open", allowFrom: ["*"] } },
@@ -3286,9 +2975,6 @@ describe("createTelegramBot", () => {
   });
 
   it("does not infer projection provenance from a current markerless outbound row", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-
     const storePath = createTelegramTestStorePath("dm-markerless-assistant");
     const config = {
       channels: { telegram: { dmPolicy: "open", allowFrom: ["*"] } },
@@ -3349,9 +3035,6 @@ describe("createTelegramBot", () => {
   });
 
   it("dedupes a Markdown assistant transcript against its complete Telegram projection", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-
     const storePath = createTelegramTestStorePath("dm-visible-dedupe");
     const config = {
       channels: {
@@ -3479,9 +3162,6 @@ describe("createTelegramBot", () => {
       ],
     },
   ])("keeps the full assistant transcript when $name is missing", async ({ parts }) => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-
     const storePath = createTelegramTestStorePath(
       `dm-incomplete-projection-${parts[0]?.messageId}`,
     );
@@ -3556,9 +3236,6 @@ describe("createTelegramBot", () => {
   });
 
   it("preserves cached multipart provenance on a markerless Telegram reply target", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-
     const storePath = createTelegramTestStorePath("dm-complete-multipart");
     const config = {
       channels: { telegram: { dmPolicy: "open", allowFrom: ["*"] } },
@@ -3645,9 +3322,6 @@ describe("createTelegramBot", () => {
   });
 
   it("keeps an assistant transcript when persisted projection metadata is invalid", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-
     const storePath = createTelegramTestStorePath("dm-invalid-projection");
     const config = {
       channels: { telegram: { dmPolicy: "open", allowFrom: ["*"] } },
@@ -3729,9 +3403,6 @@ describe("createTelegramBot", () => {
   });
 
   it("keeps identical visible replies correlated to distinct transcript identities", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-
     const storePath = createTelegramTestStorePath("dm-repeat-projection");
     const config = {
       channels: { telegram: { dmPolicy: "open", allowFrom: ["*"] } },
@@ -3822,9 +3493,6 @@ describe("createTelegramBot", () => {
   });
 
   it("does not collapse literal user Markdown into markerless plain cache text", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-
     const storePath = createTelegramTestStorePath("dm-user-markdown");
     const config = {
       channels: { telegram: { dmPolicy: "open", allowFrom: ["*"] } },
@@ -3889,9 +3557,6 @@ describe("createTelegramBot", () => {
   });
 
   it("skips direct transcript context for hard reset messages", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-
     const storePath = createTelegramTestStorePath("dm-reset-context");
     const config = {
       channels: {
@@ -3938,10 +3603,6 @@ describe("createTelegramBot", () => {
   });
 
   it("uses quote text when a Telegram partial reply is received", async () => {
-    onSpy.mockClear();
-    sendMessageSpy.mockClear();
-    replySpy.mockClear();
-
     createTelegramBot({ token: "tok" });
     const handler = getOnHandler("message") as (ctx: Record<string, unknown>) => Promise<void>;
 
@@ -3981,10 +3642,6 @@ describe("createTelegramBot", () => {
   });
 
   it("keeps reply linkage while omitting filtered binary reply captions", async () => {
-    onSpy.mockClear();
-    sendMessageSpy.mockClear();
-    replySpy.mockClear();
-
     createTelegramBot({ token: "tok" });
     const handler = getOnHandler("message") as (ctx: Record<string, unknown>) => Promise<void>;
 
@@ -4015,9 +3672,6 @@ describe("createTelegramBot", () => {
   });
 
   it("includes replied image media in inbound context for text replies", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-    getFileSpy.mockClear();
     const botShutdown = new AbortController();
     const mediaAbort = new AbortController();
     let replyGetFileSignal: AbortSignal | undefined;
@@ -4092,9 +3746,6 @@ describe("createTelegramBot", () => {
   });
 
   it("dispatches the current text when best-effort reply media times out", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-    getFileSpy.mockClear();
     const timeout = Object.assign(new Error("media response headers timed out"), {
       name: "TimeoutError",
     });
@@ -4140,9 +3791,6 @@ describe("createTelegramBot", () => {
   });
 
   it("dispatches the current text when classic polling aborts reply media", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-    getFileSpy.mockClear();
     const botShutdown = new AbortController();
     getFileSpy.mockImplementationOnce(async () => {
       botShutdown.abort();
@@ -4177,9 +3825,6 @@ describe("createTelegramBot", () => {
   });
 
   it("durably retries a spooled reply when shutdown aborts reply media", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-    getFileSpy.mockClear();
     const botShutdown = new AbortController();
     const mediaAbort = new AbortController();
     getFileSpy.mockImplementationOnce(async () => {
@@ -4225,10 +3870,6 @@ describe("createTelegramBot", () => {
   });
 
   it("hydrates reply chains from cached Telegram messages", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-    getFileSpy.mockClear();
-
     const mediaFetch = vi.fn(
       async () =>
         new Response(new Uint8Array([0x89, 0x50, 0x4e, 0x47]), {
@@ -4360,18 +4001,7 @@ describe("createTelegramBot", () => {
       chatId: 8,
     },
   ])("$name", async ({ allowFrom, expectHydrated, chatId }) => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-    getFileSpy.mockClear();
-    loadConfig.mockReturnValue({
-      channels: {
-        telegram: {
-          groupPolicy: "open",
-          contextVisibility: "allowlist_quote",
-          allowFrom,
-        },
-      },
-    });
+    mockTelegramConfig({ groupPolicy: "open", contextVisibility: "allowlist_quote", allowFrom });
 
     const mediaFetch = vi.fn(
       async () =>
@@ -4509,23 +4139,14 @@ describe("createTelegramBot", () => {
   });
 
   it("does not hydrate reply media denied by General forum topic visibility", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-    getFileSpy.mockClear();
-    loadConfig.mockReturnValue({
-      channels: {
-        telegram: {
-          groupPolicy: "allowlist",
-          contextVisibility: "allowlist",
-          groups: {
-            "-1007": {
-              requireMention: false,
-              allowFrom: ["1", "2"],
-              topics: {
-                "1": { allowFrom: ["1"], requireMention: false },
-              },
-            },
-          },
+    mockTelegramConfig({
+      groupPolicy: "allowlist",
+      contextVisibility: "allowlist",
+      groups: {
+        "-1007": {
+          requireMention: false,
+          allowFrom: ["1", "2"],
+          topics: { "1": { allowFrom: ["1"], requireMention: false } },
         },
       },
     });
@@ -4649,9 +4270,6 @@ describe("createTelegramBot", () => {
       useAccessGroup,
       expectHydrated,
     }) => {
-      onSpy.mockClear();
-      replySpy.mockClear();
-      getFileSpy.mockClear();
       const runtimeConfig = {
         ...(useAccessGroup
           ? {
@@ -4772,19 +4390,8 @@ describe("createTelegramBot", () => {
   );
 
   it("does not fetch reply media for unauthorized DM replies", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-    getFileSpy.mockClear();
-    sendMessageSpy.mockClear();
     readChannelAllowFromStore.mockResolvedValue([]);
-    loadConfig.mockReturnValue({
-      channels: {
-        telegram: {
-          dmPolicy: "pairing",
-          allowFrom: [],
-        },
-      },
-    });
+    mockTelegramConfig({ dmPolicy: "pairing", allowFrom: [] });
 
     createTelegramBot({ token: "tok" });
     const handler = getOnHandler("message") as (ctx: Record<string, unknown>) => Promise<void>;
@@ -4812,27 +4419,13 @@ describe("createTelegramBot", () => {
 
   it("defers reply media download until debounce flush", async () => {
     const DEBOUNCE_MS = 4321;
-    onSpy.mockClear();
-    replySpy.mockClear();
-    getFileSpy.mockClear();
-    loadConfig.mockReturnValue({
-      agents: {
-        defaults: {
-          envelopeTimezone: "utc",
-        },
+    mockTelegramConfig(
+      { dmPolicy: "open", allowFrom: ["*"] },
+      {
+        agents: { defaults: { envelopeTimezone: "utc" } },
+        messages: { inbound: { debounceMs: DEBOUNCE_MS } },
       },
-      messages: {
-        inbound: {
-          debounceMs: DEBOUNCE_MS,
-        },
-      },
-      channels: {
-        telegram: {
-          dmPolicy: "open",
-          allowFrom: ["*"],
-        },
-      },
-    });
+    );
 
     const mediaFetch = vi.fn(
       async () =>
@@ -4916,10 +4509,6 @@ describe("createTelegramBot", () => {
   });
 
   it("handles quote-only replies without reply metadata", async () => {
-    onSpy.mockClear();
-    sendMessageSpy.mockClear();
-    replySpy.mockClear();
-
     createTelegramBot({ token: "tok" });
     const handler = getOnHandler("message") as (ctx: Record<string, unknown>) => Promise<void>;
 
@@ -4947,10 +4536,6 @@ describe("createTelegramBot", () => {
   });
 
   it("uses top-level quote text for external partial replies", async () => {
-    onSpy.mockClear();
-    sendMessageSpy.mockClear();
-    replySpy.mockClear();
-
     createTelegramBot({ token: "tok" });
     const handler = getOnHandler("message") as (ctx: Record<string, unknown>) => Promise<void>;
 
@@ -5078,19 +4663,10 @@ describe("createTelegramBot", () => {
     onSpy.mockReset();
     sendMessageSpy.mockReset();
     replySpy.mockReset();
-    loadConfig.mockReturnValue({
-      channels: {
-        telegram: {
-          groupPolicy: "allowlist",
-          contextVisibility: "allowlist",
-          groups: {
-            "-1007": {
-              requireMention: false,
-              allowFrom: ["1"],
-            },
-          },
-        },
-      },
+    mockTelegramConfig({
+      groupPolicy: "allowlist",
+      contextVisibility: "allowlist",
+      groups: { "-1007": { requireMention: false, allowFrom: ["1"] } },
     });
 
     createTelegramBot({ token: "tok" });
@@ -5138,13 +4714,7 @@ describe("createTelegramBot", () => {
   });
 
   it("accepts group replies to the bot without explicit mention when requireMention is enabled", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-    loadConfig.mockReturnValue({
-      channels: {
-        telegram: { groups: { "*": { requireMention: true } } },
-      },
-    });
+    mockTelegramConfig({ groups: { "*": { requireMention: true } } });
 
     createTelegramBot({ token: "tok" });
     const handler = getOnHandler("message") as (ctx: Record<string, unknown>) => Promise<void>;
@@ -5170,20 +4740,12 @@ describe("createTelegramBot", () => {
   });
 
   it("prefers topic allowFrom over group allowFrom", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-    loadConfig.mockReturnValue({
-      channels: {
-        telegram: {
-          groupPolicy: "allowlist",
-          groups: {
-            "-1001234567890": {
-              allowFrom: ["123456789"],
-              topics: {
-                "99": { allowFrom: ["999999999"] },
-              },
-            },
-          },
+    mockTelegramConfig({
+      groupPolicy: "allowlist",
+      groups: {
+        "-1001234567890": {
+          allowFrom: ["123456789"],
+          topics: { "99": { allowFrom: ["999999999"] } },
         },
       },
     });
@@ -5212,20 +4774,9 @@ describe("createTelegramBot", () => {
   });
 
   it("allows group messages for per-group groupPolicy open override (global groupPolicy allowlist)", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-    loadConfig.mockReturnValue({
-      channels: {
-        telegram: {
-          groupPolicy: "allowlist",
-          groups: {
-            "-100123456789": {
-              groupPolicy: "open",
-              requireMention: false,
-            },
-          },
-        },
-      },
+    mockTelegramConfig({
+      groupPolicy: "allowlist",
+      groups: { "-100123456789": { groupPolicy: "open", requireMention: false } },
     });
     readChannelAllowFromStore.mockResolvedValueOnce(["123456789"]);
 
@@ -5247,20 +4798,9 @@ describe("createTelegramBot", () => {
   });
 
   it("blocks control commands from unauthorized senders in per-group open groups", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-    loadConfig.mockReturnValue({
-      channels: {
-        telegram: {
-          groupPolicy: "allowlist",
-          groups: {
-            "-100123456789": {
-              groupPolicy: "open",
-              requireMention: false,
-            },
-          },
-        },
-      },
+    mockTelegramConfig({
+      groupPolicy: "allowlist",
+      groups: { "-100123456789": { groupPolicy: "open", requireMention: false } },
     });
     readChannelAllowFromStore.mockResolvedValueOnce(["123456789"]);
 
@@ -5282,10 +4822,6 @@ describe("createTelegramBot", () => {
   });
 
   it("routes plugin-owned callback namespaces before synthetic command fallback", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-    editMessageTextSpy.mockClear();
-    sendMessageSpy.mockClear();
     registerPluginInteractiveHandler("codex-plugin", {
       channel: "telegram",
       namespace: "codexapp",
@@ -5311,18 +4847,15 @@ describe("createTelegramBot", () => {
     const callbackHandler = getTelegramCallbackHandlerForTests();
 
     await callbackHandler({
-      callbackQuery: {
+      callbackQuery: makeCallbackQuery({
         id: "cbq-codex-1",
         data: "codexapp:resume:thread-1",
-        from: { id: 9, first_name: "Ada", username: "ada_bot" },
         message: {
           business_connection_id: "biz-1",
-          chat: { id: 1234, type: "private" },
-          date: 1736380800,
           message_id: 11,
           text: "Select a thread",
         },
-      },
+      }),
       me: { username: "openclaw_bot" },
       getFile: getEmptyTelegramFile,
     });
@@ -5334,9 +4867,6 @@ describe("createTelegramBot", () => {
   });
 
   it("deletes plugin-owned callback messages through the bot API", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-    deleteMessageSpy.mockClear();
     registerPluginInteractiveHandler("codex-plugin", {
       channel: "telegram",
       namespace: "codexapp",
@@ -5360,17 +4890,14 @@ describe("createTelegramBot", () => {
     const callbackHandler = getTelegramCallbackHandlerForTests();
 
     await callbackHandler({
-      callbackQuery: {
+      callbackQuery: makeCallbackQuery({
         id: "cbq-codex-delete",
         data: "codexapp:delete:thread-1",
-        from: { id: 9, first_name: "Ada", username: "ada_bot" },
         message: {
-          chat: { id: 1234, type: "private" },
-          date: 1736380800,
           message_id: 11,
           text: "Select a thread",
         },
-      },
+      }),
       me: { username: "openclaw_bot" },
       getFile: getEmptyTelegramFile,
     });
@@ -5380,9 +4907,6 @@ describe("createTelegramBot", () => {
   });
 
   it("routes plugin-owned callback replies with Telegram topic params", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-    sendMessageSpy.mockClear();
     registerPluginInteractiveHandler("codex-plugin", {
       channel: "telegram",
       namespace: "codexapp",
@@ -5406,20 +4930,18 @@ describe("createTelegramBot", () => {
     const callbackHandler = getTelegramCallbackHandlerForTests();
 
     await callbackHandler({
-      callbackQuery: {
+      callbackQuery: makeCallbackQuery({
         id: "cbq-codex-topic-reply",
         data: "codexapp:reply:thread-1",
-        from: { id: 9, first_name: "Ada", username: "ada_bot" },
         message: {
           business_connection_id: "biz-topic-1",
           chat: { id: -100987654321, type: "supergroup", title: "Forum Group" },
-          date: 1736380800,
           is_topic_message: true,
           message_id: 11,
           message_thread_id: 99,
           text: "Select a thread",
         },
-      },
+      }),
       me: { username: "openclaw_bot" },
       getFile: getEmptyTelegramFile,
     });
@@ -5431,19 +4953,17 @@ describe("createTelegramBot", () => {
 
     sendMessageSpy.mockClear();
     await callbackHandler({
-      callbackQuery: {
+      callbackQuery: makeCallbackQuery({
         id: "cbq-codex-general-reply",
         data: "codexapp:reply:thread-1",
-        from: { id: 9, first_name: "Ada", username: "ada_bot" },
         message: {
           chat: { id: -100987654322, type: "supergroup", title: "Forum Group" },
-          date: 1736380800,
           is_topic_message: true,
           message_id: 12,
           message_thread_id: 1,
           text: "Select a thread",
         },
-      },
+      }),
       me: { username: "openclaw_bot" },
       getFile: getEmptyTelegramFile,
     });
@@ -5453,9 +4973,6 @@ describe("createTelegramBot", () => {
   });
 
   it("submits plugin-owned callback text through the Telegram inbound path", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-    editMessageReplyMarkupSpy.mockClear();
     const replyDone = waitForReplyCalls(1);
     registerPluginInteractiveHandler("smart-replies-plugin", {
       channel: "telegram",
@@ -5479,17 +4996,15 @@ describe("createTelegramBot", () => {
       const callbackHandler = getTelegramCallbackHandlerForTests();
 
       await callbackHandler({
-        callbackQuery: {
+        callbackQuery: makeCallbackQuery({
           id: "cbq-smart-reply-submit",
           data: "openclaw-smart-replies:v1:Rm14IGEgYnJva2VuIHRvb2w",
-          from: { id: 9, first_name: "Ada", username: "ada_bot" },
           message: {
             chat: { id: 9, type: "private" },
-            date: 1736380800,
             message_id: 11,
             text: "What should I help you sharpen next?",
           },
-        },
+        }),
         me: { username: "openclaw_bot" },
         getFile: getEmptyTelegramFile,
       });
@@ -5509,9 +5024,6 @@ describe("createTelegramBot", () => {
   });
 
   it("does not submit plugin-owned callback text when the handler declines the callback", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-    editMessageReplyMarkupSpy.mockClear();
     const handler = vi.fn(async () => ({ handled: false, submitText: "Ignore this" }));
     registerPluginInteractiveHandler("smart-replies-plugin", {
       channel: "telegram",
@@ -5535,17 +5047,15 @@ describe("createTelegramBot", () => {
       const callbackHandler = getTelegramCallbackHandlerForTests();
 
       await callbackHandler({
-        callbackQuery: {
+        callbackQuery: makeCallbackQuery({
           id: "cbq-smart-reply-declined-submit",
           data: "openclaw-smart-replies:v1:SWdub3JlIHRoaXM",
-          from: { id: 9, first_name: "Ada", username: "ada_bot" },
           message: {
             chat: { id: 9, type: "private" },
-            date: 1736380800,
             message_id: 11,
             text: "Pick a direction",
           },
-        },
+        }),
         me: { username: "openclaw_bot" },
         getFile: getEmptyTelegramFile,
       });
@@ -5562,20 +5072,13 @@ describe("createTelegramBot", () => {
   });
 
   it("does not retry plugin-owned callback text skipped by inbound policy", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-    editMessageReplyMarkupSpy.mockClear();
     const handler = vi.fn(async () => {
       // The callback was authorized before this policy change; submitText must
       // still honor the fresh inbound policy without releasing callback dedupe.
-      loadConfig.mockReturnValue({
-        channels: {
-          telegram: {
-            dmPolicy: "open",
-            allowFrom: ["*"],
-            direct: { "9": { requireTopic: true } },
-          },
-        },
+      mockTelegramConfig({
+        dmPolicy: "open",
+        allowFrom: ["*"],
+        direct: { "9": { requireTopic: true } },
       });
       return { handled: true, submitText: "Do not submit this" };
     });
@@ -5602,17 +5105,15 @@ describe("createTelegramBot", () => {
       const callbackHandler = getTelegramCallbackHandlerForTests();
 
       const callbackContext = {
-        callbackQuery: {
+        callbackQuery: makeCallbackQuery({
           id: "cbq-smart-reply-policy-skip",
           data: "openclaw-smart-replies:v1:RG8gbm90IHN1Ym1pdCB0aGlz",
-          from: { id: 9, first_name: "Ada", username: "ada_bot" },
           message: {
             chat: { id: 9, type: "private" },
-            date: 1736380800,
             message_id: 11,
             text: "Pick a direction",
           },
-        },
+        }),
         me: { username: "openclaw_bot" },
         getFile: getEmptyTelegramFile,
       };
@@ -5629,9 +5130,6 @@ describe("createTelegramBot", () => {
   });
 
   it("submits plugin-owned callback text in mention-required group topics", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-    editMessageReplyMarkupSpy.mockClear();
     const replyDone = waitForReplyCalls(1);
     registerPluginInteractiveHandler("smart-replies-plugin", {
       channel: "telegram",
@@ -5658,19 +5156,17 @@ describe("createTelegramBot", () => {
       const callbackHandler = getTelegramCallbackHandlerForTests();
 
       await callbackHandler({
-        callbackQuery: {
+        callbackQuery: makeCallbackQuery({
           id: "cbq-smart-reply-topic-submit",
           data: "openclaw-smart-replies:v1:SW52ZXN0aWdhdGUgdG9waWMgY2FsbGJhY2s",
-          from: { id: 9, first_name: "Ada", username: "ada_bot" },
           message: {
             chat: { id: -100987654321, type: "supergroup", title: "Forum Group", is_forum: true },
-            date: 1736380800,
             is_topic_message: true,
             message_id: 11,
             message_thread_id: 99,
             text: "What should I help you sharpen next?",
           },
-        },
+        }),
         me: { username: "openclaw_bot" },
         getFile: getEmptyTelegramFile,
       });
@@ -5692,9 +5188,6 @@ describe("createTelegramBot", () => {
   });
 
   it("settles spooled plugin callback text after a reply-session conflict retry succeeds", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-    editMessageReplyMarkupSpy.mockClear();
     let calls = 0;
     replySpy.mockImplementation(async (_ctx, opts) => {
       calls += 1;
@@ -5724,17 +5217,15 @@ describe("createTelegramBot", () => {
         },
       });
       const callbackHandler = getTelegramCallbackHandlerForTests();
-      const callbackQuery = {
+      const callbackQuery = makeCallbackQuery({
         id: "cbq-smart-reply-submit-retry",
         data: "openclaw-smart-replies:v1:TWFrZSBBbGljZSBmdW5uaWVy",
-        from: { id: 9, first_name: "Ada", username: "ada_bot" },
         message: {
           chat: { id: 9, type: "private" },
-          date: 1736380800,
           message_id: 11,
           text: "Pick a direction",
         },
-      };
+      });
       const update = { update_id: 403, callback_query: callbackQuery };
       const callbackContext = {
         update,
@@ -5766,9 +5257,6 @@ describe("createTelegramBot", () => {
   });
 
   it("releases plugin-owned callback dedupe when submitted text processing fails", async () => {
-    onSpy.mockClear();
-    replySpy.mockClear();
-    editMessageReplyMarkupSpy.mockClear();
     let calls = 0;
     replySpy.mockImplementation(async (_ctx, opts) => {
       calls += 1;
@@ -5801,17 +5289,15 @@ describe("createTelegramBot", () => {
       const callbackHandler = getTelegramCallbackHandlerForTests();
       const createCallbackUpdate = (updateId: number) => ({
         update_id: updateId,
-        callbackQuery: {
+        callbackQuery: makeCallbackQuery({
           id: "cbq-smart-reply-submit-fail",
           data: "openclaw-smart-replies:v1:VHJ5IHRoaXMgbGF0ZXI",
-          from: { id: 9, first_name: "Ada", username: "ada_bot" },
           message: {
             chat: { id: 9, type: "private" },
-            date: 1736380800,
             message_id: 11,
             text: "Pick a direction",
           },
-        },
+        }),
         me: { username: "openclaw_bot" },
         getFile: getEmptyTelegramFile,
       });
@@ -5847,7 +5333,6 @@ describe("createTelegramBot", () => {
       expectedAuth: true,
     },
   ])("$name", async ({ sender, messageId, expectedAuth }) => {
-    onSpy.mockClear();
     let observedAuth: TelegramInteractiveHandlerContext["auth"] | undefined;
     const handler = vi.fn(async ({ auth }: TelegramInteractiveHandlerContext) => {
       observedAuth = auth;
@@ -5897,7 +5382,6 @@ describe("createTelegramBot", () => {
   });
 
   it("passes true command auth to Telegram plugin callbacks for access-group DM senders", async () => {
-    onSpy.mockClear();
     let observedAuth: TelegramInteractiveHandlerContext["auth"] | undefined;
     const handler = vi.fn(async ({ auth }: TelegramInteractiveHandlerContext) => {
       observedAuth = auth;
@@ -5930,17 +5414,16 @@ describe("createTelegramBot", () => {
     const callbackHandler = getTelegramCallbackHandlerForTests();
 
     await callbackHandler({
-      callbackQuery: {
+      callbackQuery: makeCallbackQuery({
         id: "cbq-plugin-access-group-auth",
         data: "codexapp:resume:thread-1",
         from: { id: 123456789, first_name: "Ada", username: "ada" },
         message: {
           chat: { id: 123456789, type: "private" },
-          date: 1736380800,
           message_id: 24,
           text: "Select a thread",
         },
-      },
+      }),
       me: { username: "openclaw_bot" },
       getFile: getEmptyTelegramFile,
     });
@@ -5950,7 +5433,6 @@ describe("createTelegramBot", () => {
   });
 
   it("routes Telegram #General callback payloads as topic 1 when Telegram omits topic metadata", async () => {
-    onSpy.mockClear();
     getChatSpy.mockResolvedValue({ id: -100123456789, type: "supergroup", is_forum: true });
     const handler = vi.fn(
       async ({ respond, conversationId, threadId }: TelegramInteractiveHandlerContext) => {
@@ -5982,17 +5464,15 @@ describe("createTelegramBot", () => {
     const callbackHandler = getTelegramCallbackHandlerForTests();
 
     await callbackHandler({
-      callbackQuery: {
+      callbackQuery: makeCallbackQuery({
         id: "cbq-codex-general",
         data: "codexapp:resume:thread-1",
-        from: { id: 9, first_name: "Ada", username: "ada_bot" },
         message: {
           chat: { id: -100123456789, type: "supergroup", title: "Forum Group" },
-          date: 1736380800,
           message_id: 11,
           text: "Select a thread",
         },
-      },
+      }),
       me: { username: "openclaw_bot" },
       getFile: getEmptyTelegramFile,
     });
@@ -6029,20 +5509,9 @@ describe("createTelegramBot", () => {
       assertAuthorized: true,
     },
   ])("$name", async ({ messageThreadId, me, expectedSessionKey, assertAuthorized }) => {
-    onSpy.mockClear();
-    sendMessageSpy.mockClear();
-    commandSpy.mockClear();
-    replySpy.mockClear();
     replySpy.mockResolvedValue({ text: "response" });
 
-    loadConfig.mockReturnValue({
-      commands: { native: true },
-      channels: {
-        telegram: {
-          dmPolicy: "pairing",
-        },
-      },
-    });
+    mockTelegramConfig({ dmPolicy: "pairing" }, { commands: { native: true } });
     readChannelAllowFromStore.mockResolvedValueOnce(["12345"]);
 
     createTelegramBot({ token: "tok" });
@@ -6086,10 +5555,6 @@ describe("createTelegramBot", () => {
   });
 
   it("keeps native DM commands on the startup-resolved config when fresh reads contain SecretRefs", async () => {
-    onSpy.mockClear();
-    sendMessageSpy.mockClear();
-    commandSpy.mockClear();
-    replySpy.mockClear();
     replySpy.mockResolvedValue({ text: "response" });
 
     const startupConfig = {
@@ -6106,15 +5571,13 @@ describe("createTelegramBot", () => {
       token: "tok",
       config: startupConfig,
     });
-    loadConfig.mockReturnValue({
-      commands: { native: true },
-      channels: {
-        telegram: {
-          dmPolicy: "pairing",
-          botToken: { source: "env", provider: "default", id: "TELEGRAM_BOT_TOKEN" },
-        },
+    mockTelegramConfig(
+      {
+        dmPolicy: "pairing",
+        botToken: { source: "env", provider: "default", id: "TELEGRAM_BOT_TOKEN" },
       },
-    });
+      { commands: { native: true } },
+    );
     readChannelAllowFromStore.mockResolvedValueOnce(["12345"]);
 
     const handler = commandSpy.mock.calls.find((call) => call[0] === "status")?.[1] as
@@ -6139,19 +5602,7 @@ describe("createTelegramBot", () => {
   });
 
   it("blocks native DM commands for unpaired users", async () => {
-    onSpy.mockClear();
-    sendMessageSpy.mockClear();
-    commandSpy.mockClear();
-    replySpy.mockClear();
-
-    loadConfig.mockReturnValue({
-      commands: { native: true },
-      channels: {
-        telegram: {
-          dmPolicy: "pairing",
-        },
-      },
-    });
+    mockTelegramConfig({ dmPolicy: "pairing" }, { commands: { native: true } });
     readChannelAllowFromStore.mockResolvedValueOnce([]);
 
     createTelegramBot({ token: "tok" });
@@ -6182,7 +5633,6 @@ describe("createTelegramBot", () => {
   });
 
   it("registers message_reaction handler", () => {
-    onSpy.mockClear();
     createTelegramBot({ token: "tok" });
     const reactionHandler = onSpy.mock.calls.find((call) => call[0] === "message_reaction");
     expect(reactionHandler?.[0]).toBe("message_reaction");
@@ -6192,14 +5642,7 @@ describe("createTelegramBot", () => {
   });
 
   it("enqueues system event for reaction", async () => {
-    onSpy.mockClear();
-    enqueueSystemEventSpy.mockClear();
-
-    loadConfig.mockReturnValue({
-      channels: {
-        telegram: { dmPolicy: "open", allowFrom: ["*"], reactionNotifications: "all" },
-      },
-    });
+    mockTelegramConfig({ dmPolicy: "open", allowFrom: ["*"], reactionNotifications: "all" });
 
     createTelegramBot({ token: "tok" });
     const handler = getOnHandler("message_reaction") as (
@@ -6373,17 +5816,11 @@ describe("createTelegramBot", () => {
       expectedEnqueueCalls,
       expectedEvent,
     }) => {
-      onSpy.mockClear();
-      enqueueSystemEventSpy.mockClear();
       if (sentByBot !== undefined) {
         wasSentByBot.mockReturnValue(sentByBot);
       }
 
-      loadConfig.mockReturnValue({
-        channels: {
-          telegram: channelConfig,
-        },
-      });
+      mockTelegramConfig(channelConfig);
 
       createTelegramBot({ token: "tok" });
       const handler = getOnHandler("message_reaction") as (
@@ -6401,14 +5838,7 @@ describe("createTelegramBot", () => {
   );
 
   it("enqueues one event per added emoji reaction", async () => {
-    onSpy.mockClear();
-    enqueueSystemEventSpy.mockClear();
-
-    loadConfig.mockReturnValue({
-      channels: {
-        telegram: { dmPolicy: "open", allowFrom: ["*"], reactionNotifications: "all" },
-      },
-    });
+    mockTelegramConfig({ dmPolicy: "open", allowFrom: ["*"], reactionNotifications: "all" });
 
     createTelegramBot({ token: "tok" });
     const handler = getOnHandler("message_reaction") as (
@@ -6439,14 +5869,7 @@ describe("createTelegramBot", () => {
   });
 
   it("routes forum group reactions to the general topic (thread id not available on reactions)", async () => {
-    onSpy.mockClear();
-    enqueueSystemEventSpy.mockClear();
-
-    loadConfig.mockReturnValue({
-      channels: {
-        telegram: { dmPolicy: "open", reactionNotifications: "all" },
-      },
-    });
+    mockTelegramConfig({ dmPolicy: "open", reactionNotifications: "all" });
 
     createTelegramBot({ token: "tok" });
     const handler = getOnHandler("message_reaction") as (
@@ -6476,14 +5899,7 @@ describe("createTelegramBot", () => {
   });
 
   it("uses correct session key for forum group reactions in general topic", async () => {
-    onSpy.mockClear();
-    enqueueSystemEventSpy.mockClear();
-
-    loadConfig.mockReturnValue({
-      channels: {
-        telegram: { dmPolicy: "open", reactionNotifications: "all" },
-      },
-    });
+    mockTelegramConfig({ dmPolicy: "open", reactionNotifications: "all" });
 
     createTelegramBot({ token: "tok" });
     const handler = getOnHandler("message_reaction") as (
@@ -6510,14 +5926,7 @@ describe("createTelegramBot", () => {
   });
 
   it("uses correct session key for regular group reactions without topic", async () => {
-    onSpy.mockClear();
-    enqueueSystemEventSpy.mockClear();
-
-    loadConfig.mockReturnValue({
-      channels: {
-        telegram: { dmPolicy: "open", reactionNotifications: "all" },
-      },
-    });
+    mockTelegramConfig({ dmPolicy: "open", reactionNotifications: "all" });
 
     createTelegramBot({ token: "tok" });
     const handler = getOnHandler("message_reaction") as (
