+++ src/agents/embedded-agent-subscribe.subscribe-embedded-agent-session.block-reply-flushing.test.ts
@@ -1,5 +1,4 @@
-// Block-reply flush ordering tests ensure text buffered before tool execution
-// is delivered before the tool lifecycle event advances the turn.
+// Block-reply flush boundaries and optional callback behavior.
 import { describe, expect, it, vi } from "vitest";
 import {
   createStubSessionHarness,
@@ -310,5 +309,44 @@ describe("subscribeEmbeddedAgentSession", () => {
     };
     emit({ type: "message_end", message: transcriptOnlyMessage });
     expect(transcriptOnlyMessage).not.toHaveProperty("usage");
+  });
+});
+
+type StubSession = {
+  subscribe: (fn: (evt: unknown) => void) => () => void;
+};
+
+type SessionEventHandler = (evt: unknown) => void;
+
+describe("subscribeEmbeddedAgentSession", () => {
+  it("does not call onBlockReplyFlush when callback is not provided", () => {
+    // A missing optional flush callback should not break tool lifecycle events.
+    let handler: SessionEventHandler | undefined;
+    const session: StubSession = {
+      subscribe: (fn) => {
+        handler = fn;
+        return () => {};
+      },
+    };
+
+    const onBlockReply = vi.fn();
+
+    // No onBlockReplyFlush provided
+    subscribeEmbeddedAgentSession({
+      session: session as unknown as Parameters<typeof subscribeEmbeddedAgentSession>[0]["session"],
+      runId: "run-no-flush",
+      onBlockReply,
+      blockReplyBreak: "text_end",
+    });
+
+    // Missing onBlockReplyFlush should still accept streaming events.
+    expect(
+      handler?.({
+        type: "tool_execution_start",
+        toolName: "bash",
+        toolCallId: "tool-no-flush",
+        args: { command: "echo test" },
+      }),
+    ).toBeUndefined();
   });
 });
