+++ src/agents/embedded-agent-subscribe.subscribe-embedded-agent-session.compaction-and-tool-summaries.test.ts
@@ -1,9 +1,254 @@
-// Compaction retry subscription tests cover retry wait accounting, compaction
-// event emission, abort-on-unsubscribe, and verbose tool summary behavior.
+// Compaction retry state, fenced retry output, and tool summaries.
 import type { AssistantMessage } from "openclaw/plugin-sdk/llm";
 import { describe, expect, it, vi } from "vitest";
 import { onAgentEvent } from "../infra/agent-events.js";
 import { createSubscribedSessionHarness } from "./embedded-agent-subscribe.e2e-harness.js";
+import { subscribeEmbeddedAgentSession } from "./embedded-agent-subscribe.js";
+import { makeZeroUsageSnapshot } from "./usage.js";
+
+type SessionEventHandler = (evt: unknown) => void;
+
+describe("fenced output and compaction retries", () => {
+  it("waits for auto-compaction retry and clears buffered text", async () => {
+    // A retrying compaction invalidates any assistant text buffered from the
+    // failed attempt; waiters resolve only after the retry path reaches agent_end.
+    const listeners: SessionEventHandler[] = [];
+    const session = {
+      subscribe: (listener: SessionEventHandler) => {
+        listeners.push(listener);
+        return () => {
+          const index = listeners.indexOf(listener);
+          if (index !== -1) {
+            listeners.splice(index, 1);
+          }
+        };
+      },
+    } as unknown as Parameters<typeof subscribeEmbeddedAgentSession>[0]["session"];
+
+    const subscription = subscribeEmbeddedAgentSession({
+      session,
+      runId: "run-1",
+    });
+
+    const assistantMessage = {
+      role: "assistant",
+      content: [{ type: "text", text: "oops" }],
+    } as AssistantMessage;
+
+    for (const listener of listeners) {
+      listener({ type: "message_end", message: assistantMessage });
+    }
+
+    expect(subscription.assistantTexts.length).toBe(1);
+
+    for (const listener of listeners) {
+      listener({
+        type: "compaction_end",
+        willRetry: true,
+      });
+    }
+
+    expect(subscription.isCompacting()).toBe(true);
+    expect(subscription.assistantTexts.length).toBe(0);
+
+    let resolved = false;
+    const waitPromise = subscription.waitForCompactionRetry().then(() => {
+      resolved = true;
+    });
+
+    await Promise.resolve();
+    expect(resolved).toBe(false);
+
+    for (const listener of listeners) {
+      listener({ type: "agent_end" });
+    }
+
+    await waitPromise;
+    expect(resolved).toBe(true);
+  });
+
+  it("clears the exact usage snapshot when compaction starts a new attempt", () => {
+    const listeners: SessionEventHandler[] = [];
+    const session = {
+      subscribe: (listener: SessionEventHandler) => {
+        listeners.push(listener);
+        return () => {};
+      },
+    } as unknown as Parameters<typeof subscribeEmbeddedAgentSession>[0]["session"];
+    const subscription = subscribeEmbeddedAgentSession({ session, runId: "run-usage-reset" });
+    const assistantMessage = {
+      role: "assistant",
+      content: [{ type: "text", text: "before compaction" }],
+      usage: {
+        input: 100,
+        output: 20,
+        cacheRead: 300,
+        cacheWrite: 0,
+        totalTokens: 420,
+        cost: { input: 0, output: 0, cacheRead: 0, cacheWrite: 0, total: 0 },
+      },
+    } as AssistantMessage;
+
+    for (const listener of listeners) {
+      listener({ type: "message_end", message: assistantMessage });
+    }
+    expect(subscription.getLastAssistantUsage()).toMatchObject({
+      input: 100,
+      output: 20,
+      cacheRead: 300,
+      total: 420,
+    });
+
+    for (const listener of listeners) {
+      listener({ type: "compaction_end", willRetry: true });
+    }
+    expect(subscription.getLastAssistantUsage()).toBeUndefined();
+  });
+  it("resolves after compaction ends without retry", async () => {
+    const listeners: SessionEventHandler[] = [];
+    const session = {
+      subscribe: (listener: SessionEventHandler) => {
+        listeners.push(listener);
+        return () => {};
+      },
+    } as unknown as Parameters<typeof subscribeEmbeddedAgentSession>[0]["session"];
+
+    const subscription = subscribeEmbeddedAgentSession({
+      session,
+      runId: "run-2",
+    });
+
+    for (const listener of listeners) {
+      listener({ type: "compaction_start" });
+    }
+
+    expect(subscription.isCompacting()).toBe(true);
+
+    let resolved = false;
+    const waitPromise = subscription.waitForCompactionRetry().then(() => {
+      resolved = true;
+    });
+
+    await Promise.resolve();
+    expect(resolved).toBe(false);
+
+    for (const listener of listeners) {
+      listener({ type: "compaction_end", willRetry: false });
+    }
+
+    await waitPromise;
+    expect(resolved).toBe(true);
+    expect(subscription.isCompacting()).toBe(false);
+  });
+
+  it("resets assistant usage to a zero snapshot after compaction without retry", () => {
+    // When compaction ends the active assistant message is synthetic, so usage
+    // should reset to a zero snapshot instead of carrying stale token totals.
+    const listeners: SessionEventHandler[] = [];
+    const session = {
+      messages: [
+        {
+          role: "assistant",
+          content: [{ type: "text", text: "old" }],
+          usage: {
+            input: 120,
+            output: 30,
+            cacheRead: 5,
+            cacheWrite: 0,
+            totalTokens: 155,
+            cost: { input: 0.001, output: 0.002, cacheRead: 0, cacheWrite: 0, total: 0.003 },
+          },
+        },
+      ],
+      subscribe: (listener: SessionEventHandler) => {
+        listeners.push(listener);
+        return () => {};
+      },
+    } as unknown as Parameters<typeof subscribeEmbeddedAgentSession>[0]["session"];
+
+    subscribeEmbeddedAgentSession({
+      session,
+      runId: "run-3",
+    });
+
+    for (const listener of listeners) {
+      listener({ type: "compaction_end", willRetry: false });
+    }
+
+    const usage = (session.messages?.[0] as { usage?: unknown } | undefined)?.usage;
+    expect(usage).toEqual(makeZeroUsageSnapshot());
+  });
+});
+
+describe("canvas tool summaries", () => {
+  it("includes canvas action metadata in tool summaries", async () => {
+    // Canvas actions need their JSONL path in summaries so users can inspect the
+    // generated artifact without verbose tool output.
+    const onToolResult = vi.fn();
+
+    const toolHarness = createSubscribedSessionHarness({
+      runId: "run-canvas-tool",
+      verboseLevel: "on",
+      onToolResult,
+    });
+
+    toolHarness.emit({
+      type: "tool_execution_start",
+      toolName: "canvas",
+      toolCallId: "tool-canvas-1",
+      args: { action: "a2ui_push", jsonlPath: "/tmp/a2ui.jsonl" },
+    });
+
+    // Wait for async handler to complete
+    await Promise.resolve();
+
+    expect(onToolResult).toHaveBeenCalledTimes(1);
+    const payload = onToolResult.mock.calls.at(0)?.[0];
+    expect(payload.text).toContain("🖼️");
+    expect(payload.text).toContain("Canvas");
+    expect(payload.text).toContain("/tmp/a2ui.jsonl");
+  });
+  it("skips tool summaries when shouldEmitToolResult is false", () => {
+    const onToolResult = vi.fn();
+
+    const toolHarness = createSubscribedSessionHarness({
+      runId: "run-tool-off",
+      shouldEmitToolResult: () => false,
+      onToolResult,
+    });
+
+    toolHarness.emit({
+      type: "tool_execution_start",
+      toolName: "read",
+      toolCallId: "tool-2",
+      args: { path: "/tmp/b.txt" },
+    });
+
+    expect(onToolResult).not.toHaveBeenCalled();
+  });
+  it("emits tool summaries when shouldEmitToolResult overrides verbose", async () => {
+    const onToolResult = vi.fn();
+
+    const toolHarness = createSubscribedSessionHarness({
+      runId: "run-tool-override",
+      verboseLevel: "off",
+      shouldEmitToolResult: () => true,
+      onToolResult,
+    });
+
+    toolHarness.emit({
+      type: "tool_execution_start",
+      toolName: "read",
+      toolCallId: "tool-3",
+      args: { path: "/tmp/c.txt" },
+    });
+
+    // Wait for async handler to complete
+    await Promise.resolve();
+
+    expect(onToolResult).toHaveBeenCalledTimes(1);
+  });
+});
 
 function toolResultPayloadAt(
   onToolResult: ReturnType<typeof vi.fn>,
