--- src/agents/embedded-agent-subscribe.subscribe-embedded-agent-session.suppresses-commentary-phase-output.test.ts
@@ -1,73 +1,0 @@
-// Commentary suppression tests ensure commentary-phase assistant messages do
-// not become replies or assistantTexts before tool use.
-import type { AssistantMessage } from "openclaw/plugin-sdk/llm";
-import { describe, expect, it, vi } from "vitest";
-import { createSubscribedSessionHarness } from "./embedded-agent-subscribe.e2e-harness.js";
-
-type AssistantMessageWithPhase = AssistantMessage & {
-  // Some providers expose phase on the assistant message, while others only
-  // include it in textSignature metadata.
-  phase?: "commentary" | "final_answer";
-};
-
-describe("subscribeEmbeddedAgentSession", () => {
-  it("suppresses commentary-phase assistant messages before tool use", () => {
-    const onBlockReply = vi.fn();
-    const onPartialReply = vi.fn();
-    const { emit, subscription } = createSubscribedSessionHarness({
-      runId: "run",
-      onBlockReply,
-      onPartialReply,
-      blockReplyBreak: "message_end",
-    });
-
-    const commentaryMessage = {
-      role: "assistant",
-      phase: "commentary",
-      content: [{ type: "text", text: "Need send." }],
-      stopReason: "toolUse",
-    } as AssistantMessageWithPhase;
-
-    emit({ type: "message_start", message: commentaryMessage });
-    emit({ type: "message_end", message: commentaryMessage });
-
-    expect(onBlockReply).not.toHaveBeenCalled();
-    expect(onPartialReply).not.toHaveBeenCalled();
-    expect(subscription.assistantTexts).toStrictEqual([]);
-  });
-
-  it("suppresses commentary when phase is only present in textSignature metadata", () => {
-    const onBlockReply = vi.fn();
-    const onPartialReply = vi.fn();
-    const { emit, subscription } = createSubscribedSessionHarness({
-      runId: "run",
-      onBlockReply,
-      onPartialReply,
-      blockReplyBreak: "message_end",
-    });
-
-    const commentaryMessage = {
-      role: "assistant",
-      content: [
-        {
-          type: "text",
-          text: "Need send.",
-          textSignature: JSON.stringify({ v: 1, id: "msg_sig", phase: "commentary" }),
-        },
-      ],
-      stopReason: "toolUse",
-    } as AssistantMessage;
-
-    emit({ type: "message_start", message: commentaryMessage });
-    emit({
-      type: "message_update",
-      message: commentaryMessage,
-      assistantMessageEvent: { type: "text_delta", delta: "Need send." },
-    });
-    emit({ type: "message_end", message: commentaryMessage });
-
-    expect(onBlockReply).not.toHaveBeenCalled();
-    expect(onPartialReply).not.toHaveBeenCalled();
-    expect(subscription.assistantTexts).toStrictEqual([]);
-  });
-});
